main.js
This is the first code I have written in Javascript and spent the past couple of hours familiarizing myself with syntax. I originally thought I could access the command line variables using readline, but eventually figured out it is through process.argv . I create an array of the command line arguments and I loop through these, starting at index 2 since index 0 and 1 are the paths for the node.js and file, respectively. I chose to create two separate functions, the first of which is self explanatory, singleNumToWords. A single digit from 0-9 is passed in and the phonetic version is returned through a switch statement since there are many cases and if/else would be ugly. I also don't have to break; after each case since each case returns something. I decided to make the defualt return "NaN" just for completeness although there will never be a situation where this occurs since the value passed into this function will always be a number. For the second function, convertToWords, I have the entire number passed in and decided that the best way to get each digit is through continuous modulo by 10 and floor division by 10 to isolate each digit from each number. Each of those digits then get passed to the singleNumToWords, and I push the returned word to an array. Since this method of getting individual digits works from right to left, I concatenate the final phonetic number by looping through the array backwards so that the phonetic number is in the correct order. I return this complete phonetic number and push it to an array. I then join the array with commas and print it to console.log().