Update main.js
I realized that since the command line numbers were already being passed in as type string that I could index each number and retrieve individual digits. I got rid of the entire function where I did the mod/divide to get each digit in the previous implementation. So for each number passed in, I loop through each number's digits and call the convertToWords function with the digit casted as a number through the Number constructor. This also alleviated the reverse problem since the digits get converted to words in the correct left to right order. I also thought about my previous assumption that only numbers would be passed in as arguments and wrote an if clause to deal with a situation in which a non number argument is given by user. The conditional I used was ( ! isNan(args[index]) ) since isNan() function will return false regardless whether an actual number is of type number or string. My current functionality deals with arguments passed in that are not numbers in the following way. At line 27, if the argument is not type number, it will push the string "NotANumber" to the final array in its place. Maybe instead other specifications would require me throw an exception at this point or simply ignore non-number arguments or print something to console.log() , but since the desired output for this specific case was not specified, I didn't want to ignore this potential user error. It acknowledges the fact that there may be an error and at least lets the user know that they made an error in their input. 